Ascension and Tristan da Cunha
In 2019, the UK House of Commons Foreign Affairs Select Committee
published its report *Global Britain and the British Overseas
Territories: Resetting the relationship*, following an investigation
into the relationship between the UK and the Overseas Territories. The
report recommended that Ascension and Tristan da Cunha be recognised as
Overseas Territories in their own right (paragraph 48 / recommendation
8):

> The Committee notes that the Overseas Territory of St. Helena includes
> the separate and distinct territories of Ascension Island and Tristan
> da Cunha. These are both inhabited territories with a population that
> is not directly connected to St. Helena and have their own identities,
> elected governments and flags. Therefore, Ascension Island and Tristan
> da Cunha should be treated with equality as British Overseas
> Territories in their own right and the FCO should change their status
> to this effect. However, this change should not necessarily imply that
> the practice of St. Helena, Ascension Island and Tristan da Cunha
> sharing the same Governor should come to an
> end.^([\[68\]](https://en.wikipedia.org/wiki/Ascension_Island#cite_note-68))

To date the UK Government has not responded to the committee's
recommendation.